requirements.txt
asgiref==3.7.2
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
boto3==1.28.78
botocore==1.31.78
cffi==1.16.0
channels==4.0.0
constantly==23.10.4
cryptography==41.0.5
daphne==4.0.0
Django==4.2.7
django-ckeditor==6.7.0
django-cors-headers==4.3.0
django-filter==23.3
django-js-asset==2.1.0
django-storages==1.14.2
django-taggit==5.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
gunicorn==21.2.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
jmespath==1.0.1
Markdown==3.5.1
packaging==23.2
Pillow==10.1.0
psycopg2-binary==2.9.9
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
pymemcache==4.0.0
pyOpenSSL==23.3.0
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3.post1
s3transfer==0.7.0
service-identity==23.1.0
six==1.16.0
sqlparse==0.4.4
Twisted==23.10.0
txaio==23.1.1
typing_extensions==4.8.0
urllib3==2.0.7
zope.interface==6.1
